pyproject.toml
[project]
name = "onrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "datasets>=4.8.4",
    "fastapi>=0.136.0",
    "flashrank>=0.2.10",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.14",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "ragas>=0.3.1",
    "scipy>=1.17.1",
    "sentence-transformers>=5.4.1",
    "sounddevice>=0.5.5",
    "streamlit>=1.56.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.44.0",
]
